+++ src/components/UI/Kebab/QuestionDelete/index.jsx
@@ -1,15 +1,13 @@
-import { deleteAnswer } from 'api/answers';
-import { ReactComponent as Close } from 'assets/images/icons/ic_Close.svg';
 import PropTypes from 'prop-types';
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import ConfirmModal from 'components/ConfirmModal';
+import { deleteQuestion } from 'api/questions';
+import { ReactComponent as Close } from 'assets/images/icons/ic_Close.svg';
+import ConfirmModal from 'components/UI/Modals/ConfirmModal';
 
-const AnswerDelete = ({ id, answerId, onAnswerDeleted, onKebabClick, setIsKebabLoading, setIsToast, editId, setEditId }) => {
-  AnswerDelete.propTypes = {
+const QuestionDelete = ({ id, onDeleteQuestion, onKebabClick, setIsKebabLoading, setIsToast, editId, setEditId }) => {
+  QuestionDelete.propTypes = {
     id: PropTypes.number.isRequired,
-    answerId: PropTypes.number.isRequired,
-    onAnswerDeleted: PropTypes.func.isRequired,
+    onDeleteQuestion: PropTypes.func.isRequired,
     onKebabClick: PropTypes.func.isRequired,
     setIsKebabLoading: PropTypes.func.isRequired,
     setIsToast: PropTypes.func.isRequired,
@@ -17,8 +15,8 @@ const AnswerDelete = ({ id, answerId, onAnswerDeleted, onKebabClick, setIsKebabL
     setEditId: PropTypes.func.isRequired,
   };
 
-  const [isDeleting, setIsDeleting] = useState(false);
-  const navigate = useNavigate();
+  const [isLoading, setIsLoading] = useState(false);
+  const [error, setError] = useState(null);
 
   const [showModal, setShowModal] = useState(false);
 
@@ -32,23 +30,26 @@ const AnswerDelete = ({ id, answerId, onAnswerDeleted, onKebabClick, setIsKebabL
   };
 
   const handleModalConfirm = async () => {
-    onKebabClick(id);
-    setIsKebabLoading(true);
     setShowModal(false);
-    setIsDeleting(true);
 
     try {
-      const response = await deleteAnswer(answerId);
+      setIsKebabLoading(true);
+      setIsLoading(true);
+      setError(null);
+
+      const response = await deleteQuestion(id);
       if (!response.ok) {
-        throw new Error('답변 삭제 중 오류가 발생했습니다.');
+        throw new Error('질문 삭제 중 오류가 발생했습니다.');
       }
-      onAnswerDeleted(answerId);
-      setIsToast('답변');
+
+      setIsToast('질문');
+
+      onDeleteQuestion(id);
     } catch (err) {
-      navigate('/');
+      setError('질문 삭제 중 오류가 발생했습니다.');
     } finally {
-      setIsDeleting(false);
       setIsKebabLoading(false);
+      setIsLoading(false);
       if (editId !== null) {
         setEditId(null);
       }
@@ -57,21 +58,24 @@ const AnswerDelete = ({ id, answerId, onAnswerDeleted, onKebabClick, setIsKebabL
       }, 3000);
     }
   };
+  if (error) {
+    return <div>에러: {error}</div>;
+  }
 
   return (
     <>
       <button
         type='button'
         className='flex justify-center items-center gap-2 rounded-lg w-[103px] h-[30px] text-gray-50 hover:text-blue-50 hover:bg-gray-20'
-        disabled={isDeleting}
         onClick={handleDelete}
+        disabled={isLoading}
       >
         <Close className='w-3.5 h-3.5 fill-current' />
-        <p>답변삭제</p>
+        질문삭제
       </button>
-      <ConfirmModal isOpen={showModal} onConfirm={handleModalConfirm} onCancel={handleModalCancel} message='답변을 삭제하시겠습니까?' />
+      <ConfirmModal isOpen={showModal} onConfirm={handleModalConfirm} onCancel={handleModalCancel} message='질문을 삭제하시겠습니까?' />
     </>
   );
 };
 
-export default AnswerDelete;
+export default QuestionDelete;
